requirements.txt
'litellm[proxy]'
pyautogen
chromadb